requirements.txt
faker>=25.0
faker-commerce>=1.0
flask>=3.0
mysql-connector-python>=9.0
python-dotenv>=1.0
